Improve support for special names + limited data variable support. The parsing is a bit hacky; it shouldn't *really* separate parsing funcs from vars since it's all essentially the same thing, but with the existing design it felt easiest to 'bruteforce' it this way.
Trying to indicate 'modified types' (parsed with the ? prefix) also was a bit hacky.
Currently I only apply it to ClassType (which tracks with how it's used with return types
in function parsing), however in Wine and LLVM, it's handled far more generically.
My test cases only needed it to be used with ClassType for now, so I'll leave it at that.
I'd rather not hack with any of the types unless it's specifically needed.