Selectively store backrefs based on context. Parsers such as LLVM are able to properly use the context of what it's parsing to handle this. This is a little vague in comparison but as far general rules go, appears to work for my needs at least...
Essentially, to my understanding, backrefs should be made to type or template names.
Symbol names (such as a method/function's name) should be excluded. Since these are parsed first, we essentially selectively
backref the first name or not -- depending on the flag we provided _p_qname(), which depends on its caller.

Types (_p_type()) will backref the first name, while the root node will primarily choose to not (where templates are involved, this will be overridden).

For example:
std::set<int>::erase() - first name ('erase') should NOT have a backref because it is a method name.
std::set<int> - first name ('set<int>') SHOULD have a backref because it is a type.
std::set<int>::templated_method<T>() - first name ('templated_method<T>()') SHOULD have a backref because it is a templated name; the template args can refer back to it.
I'm not sure if this last one is always the case, but it seems to be at least some of the time - we'll assume it's always the case for now.